refactor: logic into container
